Move `refurb/helpers` utils to `ruff_python_semantic` for broader use
