+++ crates/ruff/src/rules/refurb/mod.rs
@@ -1,6 +1,5 @@
 //! Rules from [refurb](https://pypi.org/project/refurb/)/
 
-mod helpers;
 pub(crate) mod rules;
 
 #[cfg(test)]
